+++ woocommerce-heidelpay.php
@@ -10,7 +10,7 @@ use Heidelpay\MessageCodeMapper\MessageCodeMapper;
  * Plugin Name: heidelpay WooCommerce
  * Plugin URI: https://dev.heidelpay.com
  * Description: heidelpay payment integration for WooCommerce
- * Version: 1.4.0
+ * Version: 1.5.0
  * Author: heidelpay
  * Author URI: htts://www.heidelpay.com
  * Developer: heidelpay
@@ -29,7 +29,7 @@ if (in_array('woocommerce/woocommerce.php', apply_filters('active_plugins', get_
     /**
      * Required minimums and constants
      */
-    define('WC_HEIDELPAY_VERSION', '1.4.0');
+    define('WC_HEIDELPAY_VERSION', '1.5.0');
     define('WC_HEIDELPAY_MIN_PHP_VER', '5.6.0');
     define('WC_HEIDELPAY_MIN_WC_VER', '3.0.0');
     define('WC_HEIDELPAY_MAIN_FILE', __FILE__);
@@ -134,6 +134,7 @@ if (in_array('woocommerce/woocommerce.php', apply_filters('active_plugins', get_
                 require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-idl.php';
                 require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-dd.php';
                 require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-gp.php';
+                require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-iv.php';
                 require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-ivpg.php';
                 require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-so.php';
                 require_once __DIR__ . '/includes/gateways/class-wc-heidelpay-gateway-va.php';
@@ -155,6 +156,7 @@ if (in_array('woocommerce/woocommerce.php', apply_filters('active_plugins', get_
                 $methods[] = 'WC_Gateway_HP_IDL';
                 $methods[] = 'WC_Gateway_HP_DD';
                 $methods[] = 'WC_Gateway_HP_GP';
+                $methods[] = 'WC_Gateway_HP_IV';
                 $methods[] = 'WC_Gateway_HP_IVPG';
                 $methods[] = 'WC_Gateway_HP_SO';
                 $methods[] = 'WC_Gateway_HP_VA';